Write log messages directly to zLib stream
Replaces printf-like output with direct writing.
gzprintf function is limited by the zLib buffer and some long log
messages may not fit into this limit, which leads to buffer overflow
with unpredictable consequences.

In addition, now the first log message uses the current locale to fix
the start date of the current log.

Change-Id: I24b8cb7849d5025c8b46f54d55723bb1673561cf
Signed-off-by: Artem Senichev <a.senichev@yadro.com>