Update patch set 2
Patch Set 2:

The main goal of this project was to collect the latest messages from the host, because it takes a lot of memory to store all the output.
I think you don't need to modify the hostlogger to send data to journal, you can use systemd-cat for that, just connect it with the obmc-console.log file.

Patch-set: 2
Reviewer: Gerrit User 1000131 <1000131@adbd0a64-1f21-4d83-b585-671fe73cb6e4>