Update patch set 2
Patch Set 2:

(3 comments)

Excuse me guys, but I still don't understand why you want to implement this feature in the phosphor-hostlogger project. This service is not intended to be a "stream forwarder", it is a buffer for logs, it collect host output and flush it according to the configured policy. This policy is the largest part of the module (configs, dbus watchers, etc), but you don't need all of these things for your purposes. You don't even need a line ending handler (it prevents printing the message until '\n' which is not acceptable in your case).
I believe the real time streaming should be implemented either in a standalone module or in lower layer (in obmc-console-server).

Patch-set: 2
Reviewer: Gerrit User 1000192 <1000192@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Reviewer: Gerrit User 1000131 <1000131@adbd0a64-1f21-4d83-b585-671fe73cb6e4>