Update patch set 6
Patch Set 6:

(3 comments)

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > Hi Artem, any new feedbacks?
> 
> Now we do not even have a common name =)
> I believe there will be less in common with each iteration. 
> 
> Sorry, Nan, I still think that it would be better to have different repositories for different application.

While I think there's some cleanups to be done (ie, this shouldn't be a totally separate app, but just a configure mode in the existing app) I like that a lot of the code is shared.  This is a good thing overall, and moving to a different repository would necessitate copying and pasting a lot of code, which is an anti-pattern.

I haven't reviewed the code in depth, but I still think this is the right approach for the community, as it reduces the amount of code overall, and ensures that we have one application that supports everyone's use cases.  I don't see how we can get that with a separate repository.

Patch-set: 6
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>