Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> Nan, I as well think that you need to consider adding this functionality directly into obmc-console-server or indeed making a separate project.

Thanks for your feedback. I think this stream application is just another a client of obmc-console-server. It might not make sense to make a separate path just for DGRAM packets in obmc-console-server.

The main reason we add codes in HostLogger is to reuse codes and make this repo a central place for people who need serial logs.

Patch-set: 6