Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> Nan, I as well think that you need to consider adding this functionality directly into obmc-console-server or indeed making a separate project.

So unfortunately I didn't get a chance to comment on this when it was discussed on the mailing list, and so this might rehash some of the email convos. As I see it, there are a couple of different layers/angles that this discussion can look at.

I think everyone agrees that we should be trying to maximize code reuse, but there is a legitimate question on whether these applications and use cases will converge or diverge over time.

To that end, it's worth thinking about the community plans to support serial console log extraction. Right now I really see three big options:
  1) Create an ad-hoc protocol for streaming the console.
  2) Package logs into files and extract using scp, ftp, etc.
  3) Use a standard protocol like Redfish (or for the masochists among us, IPMI)

In past we've used option (1), and I expect a lot of people go down that road as well. However, I think that there is some space for a standards based solution here. And if we had a standard, I think the Redfish LogService is probably the lowest hanging fruit here.

So that leads to the next question, where should the log entries be stored for the Redfish service. I think there are really three answers here:
  1) Have bmcweb subscribe to the console stream and manage/buffer entries
  2) Keep bmcweb stateless, and have an intermediate process (such as obmc-console) manage the log entries.
  3) Merge the entry management with the entry streaming in phosphor-hostlogger.

Now option (1) is not really feasible, since as designed bmcweb really should be staying out of state management. So it's really a question of (2) or (3).  Quite frankly option 2 looks like a really clunky state. Not only would we have several layers of indirection, but we'd also have two daemons that were managing their own log rotation.

This feels like a recipe for fragmentation. As it stands I am already really concerned with how fragmented log management is in OpenBMC today. In a way keeping these things in the same repo is a way to keep the door open for unification later. 

In my opinion the differences in use cases here are not unbridgeable, and we can probably streamline this over time. Maybe it takes a while to sort out, but convergence remains an option. If we split into separate repos that tacitly says that divergence is inevitable.

Patch-set: 6
CC: Gerrit User 1000523 <1000523@adbd0a64-1f21-4d83-b585-671fe73cb6e4>