Update patch set 6
Patch Set 6:

> 1. There will be a single application (single binary), a single config, a single systemd unit, and a single recipe;
Great!

> 2. I would still like to keep compile time flags to choose whether to build both modes; 
It's not necessary, but I don't mind.

> 3. I will add support to enable/disable modes at runtime: 
>   a. register a property (something like mode_enabled) for the each mode (the buffer mode and the stream mode); 
>   b. add a match to watch the property change;
>   c. if the property changes from enabled to disabled, dynamically mark corresponding sd_event_sources (for buffer mode, event sources will be io & signal; for stream mode, event sources will be just io) as disabled and remove matches if any (the buffer based service has a match for host states). The other ways work similarly, we can disable a service at runtime;
When I wrote about switching modes, I meant control through the settings file /etc/hostlogger/*.conf.
Configuring a service is a very rare operation, we do not need to have variables in the DBus for this. 
It is enough to set up and start the corresponding modes in main(). 

> To implement such proposal, I plan to refactor some codes (service.cpp and dbusloop.cpp) and introduce sdbusplus + sdeventplus (the two C++ wrappers for libsystemd; I think they are quite common in OpenBMC codes) for shorter and cleaner codes.
I deliberately gave up using the sdeventplus library a year ago. At that time, its functionality was not enough and I had to write my own handlers.
I would not rewrite this part of the code now, it is better to do such refactoring in a separate task.

Patch-set: 6
Reviewer: Gerrit User 1000131 <1000131@adbd0a64-1f21-4d83-b585-671fe73cb6e4>